requirements.txt
asgiref==3.8.1
Django==5.1.4
django-debug-toolbar==4.4.6
Faker==33.1.0
pillow==11.2.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2025.2
django-widget-tweaks
psycopg2-binary==2.9.10
python-decouple==3.8